Moved printings state to decks.svelte.ts
